Fix CoordTest
